requirements.txt
discord.py==2.3.2
aiohttp==3.9.3
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
pydantic==2.6.1
pytest==8.0.0
pytest-asyncio==0.23.5
black==24.1.1
flake8==7.0.0
mypy==1.8.0